requirements.txt
allure-pytest==2.13.5
Flask==3.1.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-rerunfailures==15.0
requests==2.32.3
selenium==4.29.0
waitress==3.0.2
pyserial==3.5
